+++ changelog.md
@@ -2,6 +2,7 @@
 
 * Add `fromListWithDef`, `fromVectorWithDef`, `interleave`.
 * Add `unfoldr` and `unfoldrM`.
+* Export `tabulateFixM'`.
 * Add `sliceSubvectors`, `traverseSubvectors`, `zipWithSubvectors` and `zipWithMSubvectors`.
 * Deprecate `zipSubvectors` in favor of `zipWithSubvectors`.